removed unused imports of deprecated classes, fixed javadoc errors
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@764203 13f79535-47bb-0310-9956-ffa450edef68

+++ src/contrib/src/org/apache/poi/hssf/contrib/view/SVTableModel.java
@@ -61,7 +61,7 @@ public class SVTableModel extends AbstractTableModel {
     HSSFRow r = st.getRow(row);
     HSSFCell c = null;
     if (r != null) {
-      c = r.getCell((short)col);
+      c = r.getCell(col);
     }
     return c;
   }
